% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avsSize.R
\name{avsSize}
\alias{avsSize}
\title{avsSize}
\usage{
avsSize(avs)
}
\arguments{
\item{avs}{A GRanges object which is outputted by loadLd function}
}
\description{
This function will output a dataframe of LD size of each tag SNP
}
\examples{
ld<-loadLd(file.path(system.file("extdata", "ld_BCa_raggr.csv", package="VSE")), type="raggr")
avs<-makeAVS(ld)
avsSize(avs)
}
\keyword{AVS,Granges}

