% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedToGRanges.R
\name{bedToGRanges}
\alias{bedToGRanges}
\title{bedToGRanges}
\usage{
bedToGRanges(file)
}
\arguments{
\item{file}{A bed file. Must contain at least three columns: chr, start and end.}
}
\description{
This function will convert a bed to GRanges object
}
\examples{
\dontrun{
bedToGRanges(file)
}
}
\keyword{bed,Granges}

