% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectMatrix.R
\name{intersectMatrix}
\alias{intersectMatrix}
\title{intersectMatrix}
\usage{
intersectMatrix(avs, regions, ...)
}
\arguments{
\item{avs}{A GRanges object which is outputted by loadLd function}

\item{regions}{A data frame. The data frame contains sample sheet identical to DiffBind or ChIPQC input sample sheets}

\item{...}{Heatmap arguments}
}
\description{
This function will count the intersection tally of AVS and genomic regions
}
\examples{
\dontrun{
intersectMatrix(avs,
                regions=samples,
                col=c("white","grey10"),
                scale="none",
                margins=c(10,5),
                cexRow = 1,
                cexCol = 0.5,
                Rowv=NA,
                Colv=NA)
}
}
\keyword{AVS,Granges}

