% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSampleRegions.R
\name{loadSampleRegions}
\alias{loadSampleRegions}
\title{loadSampleRegions}
\usage{
loadSampleRegions()
}
\value{
A directory names VSE_samples that will contain 6 bed files and 1 sampleSheet.csv
}
\description{
The sample bed files are DNAse-seq, ChIP-seq for H3K4me1, H3K4me3, H3K27ac, H3K27me3 and H3K36me3 for MCF7 cells. The data are obtained from ENCODE project. There is also one sampleSheet.csv which is the sample sheet for the bed regions in the format similar to ChIPQC or DiffBind requirement.
}
\details{
This function will download sample bed files from www.hansenhelab.org/VSE/sample_regions in /VSE_samples
}
\examples{
\dontrun{
loadSampleRegions()
}
}
\keyword{sample,histone,MCF7}

