% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAvs.R
\name{makeAVS}
\alias{makeAVS}
\title{makeAVS}
\usage{
makeAVS(x)
}
\arguments{
\item{x}{A GRanges object which is outputted by loadLd function}
}
\description{
This function will create disjoint LD blocks from a GRanges object
}
\examples{
ld<-loadLd(file.path(system.file("extdata", "ld_BCa_raggr.csv", package="VSE")), type="raggr")
avs<-makeAVS(ld)
}
\keyword{AVS,Granges}

