% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantSetEnrichment.R
\name{variantSetEnrichment}
\alias{variantSetEnrichment}
\title{variantSetEnrichment}
\usage{
variantSetEnrichment(avs, mrvs, regions)
}
\arguments{
\item{avs}{A GRangesList object which is outputted by makeAVS function}

\item{mrvs}{A list of GRangesList objects outputted by makeMRVS function}

\item{regions}{A data frame. The data frame contains sample sheet identical to DiffBind or ChIPQC input sample sheets}
}
\description{
This function will calculate matching random variant sets (MRVS) idential to AVS
}
\examples{
\dontrun{
variantSetEnrichment(avs, mrvs, regions=samples)
#We have included the output from our example analysis with the package. \\
You can load the example VariantSetEnrichment output by typing:
load(file.path(system.file("extdata", "vse_output.Rda", package="VSE")))
}
}
\keyword{VSE}

