% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSEplot.R
\name{VSEplot}
\alias{VSEplot}
\title{VSEplot}
\usage{
VSEplot(data, padj = 0.01, ...)
}
\arguments{
\item{data}{A list of matrices outputted by the function VariantSetEnrichment}

\item{padj}{Bonferroni adjusted p-value cutoff. Default: 0.01}

\item{...}{Arguments from boxplot}
}
\description{
This function will generate a figure for VSE analysis
}
\examples{
#Load pre-saved object "bca.vse" as an example VSE output
load(file.path(system.file("extdata", "vse_output.Rda", package="VSE")))
VSEplot(bca.vse, las=2,pch=20, cex=1, cex.main=0.6, padj=0.05)
}
\keyword{VSE,}
\keyword{boxplot}

