% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSEqq.R
\name{VSEqq}
\alias{VSEqq}
\title{VSEqq}
\usage{
VSEqq(data, ...)
}
\arguments{
\item{data}{A matrix outputted by the function VariantSetEnrichment}

\item{...}{Arguments for qqnorm plot}
}
\description{
This function will generate QQ plots of normal distribution
}
\examples{
#Load pre-saved object "bca.vse" as an example VSE output
load(file.path(system.file("extdata", "vse_output.Rda", package="VSE")))
VSEqq(bca.vse)
}
\keyword{VSE,}
\keyword{boxplot}

