% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSESummary.R
\name{VSESummary}
\alias{VSESummary}
\title{VSESummary}
\usage{
VSESummary(data)
}
\arguments{
\item{data}{A matrix outputted by the function VariantSetEnrichment}
}
\description{
This function will compute the enrichment from a VSE matrix
}
\examples{
#Load pre-saved object "bca.vse" as an example VSE output
load(file.path(system.file("extdata", "vse_output.Rda", package="VSE")))
VSESummary(bca.vse)
}
\keyword{VSE,}
\keyword{enrichment}

